+++ packages/genre-tree-view/src/GenreTreeWheelRadialPopCoreBase.tsx
@@ -445,10 +445,7 @@ export function WheelRadialPopCoreCore({
           playState,
         },
         wheelItemCountRange,
-        undefined,
-        undefined,
-        coreRootCircleRadius,
-        rootLinkOrigin,
+        { radialReferenceRadius: coreRootCircleRadius, rootLinkOrigin },
       );
     });
 
@@ -496,10 +493,7 @@ export function WheelRadialPopCoreCore({
           playState,
         },
         wheelItemCountRange,
-        false,
-        true,
-        coreRootCircleRadius,
-        rootLinkOrigin,
+        { isCoreSector: true, radialReferenceRadius: coreRootCircleRadius, rootLinkOrigin },
       );
     });
 
@@ -541,9 +535,7 @@ export function WheelRadialPopCoreCore({
           playState,
         },
         wheelItemCountRange,
-        true,
-        undefined,
-        coreRootCircleRadius,
+        { skipRootNode: true, radialReferenceRadius: coreRootCircleRadius, isMainstreamSector: true },
       );
     }
   }, [